requirements.txt
google-generativeai
arxiv
requests
jupyter
notebook
Flask